+++ js/plugins/table.js
@@ -1,4 +1,4 @@
-var TableConfig = {
+var TableDefaultConfig = {
     templateBeginToken: "<%",
     templateEndToken: "%>",
     paginationDistance: 5,
@@ -119,7 +119,7 @@ var TableConfig = {
 };
 
 Metro.tableSetup = function(options){
-    TableConfig = $.extend({}, TableConfig, options);
+    TableDefaultConfig = $.extend({}, TableDefaultConfig, options);
 };
 
 if (typeof window.metroTableSetup !== undefined) {
@@ -128,8 +128,7 @@ if (typeof window.metroTableSetup !== undefined) {
 
 var Table = {
     init: function( options, elem ) {
-        this.options = TableConfig;
-        this.options = $.extend( {}, this.options, options );
+        this.options = $.extend( {}, TableDefaultConfig, options );
         this.elem  = elem;
         this.element = $(elem);
         this.currentPage = 1;
@@ -172,127 +171,6 @@ var Table = {
 
         return this;
     },
-
-    //options: {
-
-        // templateBeginToken: "<%",
-        // templateEndToken: "%>",
-        // paginationDistance: 5,
-        //
-        // locale: METRO_LOCALE,
-        //
-        // horizontalScroll: false,
-        // horizontalScrollStop: null,
-        // check: false,
-        // checkType: "checkbox",
-        // checkStyle: 1,
-        // checkColIndex: 0,
-        // checkName: null,
-        // checkStoreKey: "TABLE:$1:KEYS",
-        // rownum: false,
-        // rownumTitle: "#",
-        //
-        // filters: null,
-        // filtersOperator: "and",
-        //
-        // source: null,
-        //
-        // searchMinLength: 1,
-        // searchThreshold: 500,
-        // searchFields: null,
-        //
-        // showRowsSteps: true,
-        // showSearch: true,
-        // showTableInfo: true,
-        // showPagination: true,
-        // paginationShortMode: true,
-        // showActivity: true,
-        // muteTable: true,
-        //
-        // rows: 10,
-        // rowsSteps: "10,25,50,100",
-        //
-        // staticView: false,
-        // viewSaveMode: "client",
-        // viewSavePath: "TABLE:$1:OPTIONS",
-        //
-        // sortDir: "asc",
-        // decimalSeparator: ".",
-        // thousandSeparator: ",",
-        //
-        // tableRowsCountTitle: "Show entries:",
-        // tableSearchTitle: "Search:",
-        // tableInfoTitle: "Showing $1 to $2 of $3 entries",
-        // paginationPrevTitle: "Prev",
-        // paginationNextTitle: "Next",
-        // allRecordsTitle: "All",
-        // inspectorTitle: "Inspector",
-        //
-        // activityType: "cycle",
-        // activityStyle: "color",
-        // activityTimeout: 100,
-        //
-        // searchWrapper: null,
-        // rowsWrapper: null,
-        // infoWrapper: null,
-        // paginationWrapper: null,
-        //
-        // cellWrapper: true,
-        //
-        // clsComponent: "",
-        // clsTableContainer: "",
-        // clsTable: "",
-        //
-        // clsHead: "",
-        // clsHeadRow: "",
-        // clsHeadCell: "",
-        //
-        // clsBody: "",
-        // clsBodyRow: "",
-        // clsBodyCell: "",
-        // clsCellWrapper: "",
-        //
-        // clsFooter: "",
-        // clsFooterRow: "",
-        // clsFooterCell: "",
-        //
-        // clsTableTop: "",
-        // clsRowsCount: "",
-        // clsSearch: "",
-        //
-        // clsTableBottom: "",
-        // clsTableInfo: "",
-        // clsTablePagination: "",
-        //
-        // clsPagination: "",
-        //
-        // clsEvenRow: "",
-        // clsOddRow: "",
-        // clsRow: "",
-        //
-        // onDraw: Metro.noop,
-        // onDrawRow: Metro.noop,
-        // onDrawCell: Metro.noop,
-        // onAppendRow: Metro.noop,
-        // onAppendCell: Metro.noop,
-        // onSortStart: Metro.noop,
-        // onSortStop: Metro.noop,
-        // onSortItemSwitch: Metro.noop,
-        // onSearch: Metro.noop,
-        // onRowsCountChange: Metro.noop,
-        // onDataLoad: Metro.noop,
-        // onDataLoadError: Metro.noop,
-        // onDataLoaded: Metro.noop,
-        // onFilterRowAccepted: Metro.noop,
-        // onFilterRowDeclined: Metro.noop,
-        // onCheckClick: Metro.noop,
-        // onCheckClickAll: Metro.noop,
-        // onCheckDraw: Metro.noop,
-        // onViewSave: Metro.noop,
-        // onViewGet: Metro.noop,
-        // onViewCreated: Metro.noop,
-        // onTableCreate: Metro.noop
-    //},
 
     _setOptionsFromDOM: function(){
         var element = this.element, o = this.options;